Reorganize koulutus page files
